requirements.txt
# === 基础图像处理 ===
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0

# === 报告与 API 服务 ===
openpyxl>=3.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
httpx>=0.25.0

# === 深度学习推理 (YOLO26x) ===
ultralytics>=8.3.0
torch>=2.4.0
torchvision>=0.19.0
# transformers is optional (used for fallback OCR)
transformers>=4.30.0
